Reworking cli structure, breaking out execution.  Making FsUtil a shell, and enforcing we rely on @travetto/base when possible.
