+++ module/base/src/bootstrap/fs-util.js
@@ -60,7 +60,7 @@ const FsUtil = {
 
     return `${fileContents};\nexport const _$TRV = 1;`;
   },
-  resolveFrameworkFile: (pth) => {
+  resolveFrameworkDevFile: (pth) => {
     if (pth.includes('@travetto')) {
       pth = FsUtil.toUnix(pth).replace(/.*\/@travetto\/([^/]+)\/([^@]+)$/g, (all, name, rest) => {
         const mid = subName === name ? '' : `node_modules/@travetto/${name}/`;
