+++ module/cli/src/util.js
@@ -1,11 +1,8 @@
 //@ts-check
 const commander = require('commander');
-const fs = require('fs');
 const child_process = require('child_process');
 
 const { FsUtil } = require('./fs-util');
-
-const PREFIX = 'travetto-cli';
 
 const COLORS = {
   blue: `\x1b[94m`,
@@ -19,10 +16,6 @@ const COLORS = {
   reset: `\x1b[0m`
 };
 
-commander
-  // @ts-ignore
-  .version(require('../package.json').version);
-
 const Util = {
   program: commander,
   dependOn(cmd, args, s_cwd) {
@@ -32,48 +25,6 @@ const Util = {
       stdio: [0, 1, 2],
       shell: true
     });
-  },
-  requireModule(f) {
-    let p = FsUtil.toUnix(fs.realpathSync(`${FsUtil.cwd}/node_modules/.bin/${f}`));
-    if (!p.startsWith(FsUtil.cwd)) {
-      p = `${FsUtil.cwd}/node_modules/@travetto/${p.split('travetto/module/')[1]}`;
-    }
-    require(p).init();
-  },
-  loadAllPlugins() {
-    const BIN_DIR = `${FsUtil.cwd}/node_modules/.bin`;
-    if (fs.existsSync(BIN_DIR)) {
-      const files = fs.readdirSync(BIN_DIR).filter(x => x.startsWith(`${PREFIX}-`));
-      for (const f of files) {
-        this.requireModule(f);
-      }
-    }
-  },
-  loadSinglePlugin(cmd) {
-    try {
-      this.requireModule(`${PREFIX}-${cmd.replace(/:.*$/, '')}`);
-    } catch (e) {
-      console.error('Unknown command', cmd);
-      this.showHelp(1);
-    }
-  },
-  showHelp(code = 0) {
-    this.loadAllPlugins();
-    commander.help();
-    process.exit(code);
-  },
-  execute(args) {
-    const cmd = args[2];
-
-    if (!cmd) {
-      this.showHelp();
-    } else if (!cmd.startsWith('-')) {
-      this.loadSinglePlugin(cmd);
-    } else {
-      this.loadAllPlugins();
-    }
-
-    commander.parse(args);
   },
   fork(cmd, args, env) {
     return new Promise((resolve, reject) => {
